refactor: move event handlers to separate folder
