+++ src/handlers/index.ts
@@ -1,0 +1,9 @@
+import { onGuildMemberUpdate } from './guildMemberUpdate';
+import { onGuildMemberAdd } from './guildMemberAdd';
+import { onGuildMemberRemove } from './guildMemberRemove';
+
+export {
+    onGuildMemberAdd,
+    onGuildMemberUpdate,
+    onGuildMemberRemove
+}
